Add helper file for chromacoded disassembly; add checkmark for whether branch will be taken
